Backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
bcrypt==4.0.1
pyjwt==2.8.0
numpy==1.26.0
scikit-learn==1.3.0
joblib==1.3.2
gunicorn==21.2.0